+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Change Log
 
 - [beta] Fix an issue where the current line highlight wrongly shifted when the overscrolling is set.
 - [beta] Fix an issue where line numbers at the bottm part were hidden when the overscrolling is set.
+- [beta] Fix an issue where syntax highlight was ocasionally not updated when the text is edited.
 - [beta] Fix an issue where the current syntax highlight was not removed when selecting style “None.”
 - [beta] Fix an issue where the visibility of invisibles of opened documents cannot be changed.
 - [beta] Fix an issue where the previous current line highligts could remain.